requirements.txt
python == 3.10
Django == 4.0
djangorestframework == 3.12.4
djoser == 2.1.0
djangorestframework_simplejwt == 4.8.0
django-filter == 21.1
requests == 2.26.0
PyJWT == 2.1.0